fix typing of dragdroppable
