+++ superset-frontend/src/dashboard/types.ts
@@ -139,6 +139,12 @@ export type LayoutItem = {
   };
 };
 
+export type LayoutItemWithCustomMarkdown = LayoutItem & {
+  meta: LayoutItem['meta'] & {
+    customMarkdown: CustomMarkdown;
+  };
+};
+
 type ActiveFilter = {
   scope: number[];
   values: ExtraFormData;
@@ -160,6 +166,27 @@ export type EmbeddedDashboard = {
   uuid: string;
   dashboard_id: string;
   allowed_domains: string[];
+};
+
+export type CustomMarkdown = {
+  custom_markdown_id: string;
+  name: string;
+  project_id: string;
+  definition: any;
+  directory: CustomMarkdownDirectory;
+  created_at: string;
+  modified_at: string;
+};
+
+export type CustomMarkdowns = CustomMarkdown[];
+
+export type CustomMarkdownDirectory = {
+  directory_id: string;
+  name: string;
+  type: 'CUSTOM_MARKDOWN';
+  project_id: string;
+  parent_id: string;
+  size: string;
 };
 
 export type Orientation = typeof orientations[keyof typeof orientations];
